refactor: changed acceleration type from int to float

+++ src/protobuf/mmc/command.pb.zig
@@ -617,13 +617,13 @@ pub const Request = struct {
         pub const Line = struct {
             line: u32 = 0,
             velocity: ?f32 = null,
-            acceleration: ?u32 = null,
+            acceleration: ?f32 = null,
             velocity_mode: Request.VelocityMode = @enumFromInt(0),
 
             pub const _desc_table = .{
                 .line = fd(1, .{ .scalar = .uint32 }),
                 .velocity = fd(2, .{ .scalar = .float }),
-                .acceleration = fd(3, .{ .scalar = .uint32 }),
+                .acceleration = fd(3, .{ .scalar = .float }),
                 .velocity_mode = fd(4, .@"enum"),
             };
 
@@ -743,7 +743,7 @@ pub const Request = struct {
         line: u32 = 0,
         carrier: u32 = 0,
         velocity: f32 = 0,
-        acceleration: u32 = 0,
+        acceleration: f32 = 0,
         control: mmc.Control = @enumFromInt(0),
         disable_cas: bool = false,
         velocity_mode: Request.VelocityMode = @enumFromInt(0),
@@ -769,7 +769,7 @@ pub const Request = struct {
             .line = fd(1, .{ .scalar = .uint32 }),
             .carrier = fd(2, .{ .scalar = .uint32 }),
             .velocity = fd(3, .{ .scalar = .float }),
-            .acceleration = fd(4, .{ .scalar = .uint32 }),
+            .acceleration = fd(4, .{ .scalar = .float }),
             .control = fd(8, .@"enum"),
             .disable_cas = fd(9, .{ .scalar = .bool }),
             .velocity_mode = fd(10, .@"enum"),
@@ -837,7 +837,7 @@ pub const Request = struct {
         axis: u32 = 0,
         direction: Request.Direction = @enumFromInt(0),
         velocity: f32 = 0,
-        acceleration: u32 = 0,
+        acceleration: f32 = 0,
         carrier: ?u32 = null,
         velocity_mode: Request.VelocityMode = @enumFromInt(0),
 
@@ -846,7 +846,7 @@ pub const Request = struct {
             .axis = fd(2, .{ .scalar = .uint32 }),
             .direction = fd(3, .@"enum"),
             .velocity = fd(4, .{ .scalar = .float }),
-            .acceleration = fd(5, .{ .scalar = .uint32 }),
+            .acceleration = fd(5, .{ .scalar = .float }),
             .carrier = fd(6, .{ .scalar = .uint32 }),
             .velocity_mode = fd(7, .@"enum"),
         };
@@ -913,7 +913,7 @@ pub const Request = struct {
         carrier: u32 = 0,
         direction: Request.Direction = @enumFromInt(0),
         velocity: f32 = 0,
-        acceleration: u32 = 0,
+        acceleration: f32 = 0,
         transition: ?Request.Pull.Transition = null,
         velocity_mode: Request.VelocityMode = @enumFromInt(0),
 
@@ -923,7 +923,7 @@ pub const Request = struct {
             .carrier = fd(3, .{ .scalar = .uint32 }),
             .direction = fd(4, .@"enum"),
             .velocity = fd(5, .{ .scalar = .float }),
-            .acceleration = fd(6, .{ .scalar = .uint32 }),
+            .acceleration = fd(6, .{ .scalar = .float }),
             .transition = fd(7, .submessage),
             .velocity_mode = fd(8, .@"enum"),
         };
